+++ allennlp/training/metrics/boolean_accuracy.py
@@ -3,6 +3,7 @@ from typing import Optional
 from overrides import overrides
 import torch
 
+from allennlp.nn.util import FloatPrecision
 from allennlp.training.metrics.metric import Metric
 
 
@@ -51,9 +52,9 @@ class BooleanAccuracy(Metric):
 
             # We want to skip predictions that are completely masked;
             # so we'll keep predictions that aren't.
-            keep = mask.view(batch_size, -1).max(dim=1)[0].float()
+            keep = mask.view(batch_size, -1).max(dim=1)[0].type(FloatPrecision.dtype)
         else:
-            keep = torch.ones(batch_size).float()
+            keep = torch.ones(batch_size).type(FloatPrecision.dtype)
 
         predictions = predictions.view(batch_size, -1)
         gold_labels = gold_labels.view(batch_size, -1)
@@ -62,7 +63,7 @@ class BooleanAccuracy(Metric):
         # so .eq -> .prod will be 1 if every element of the instance prediction is correct
         # and 0 if at least one element of the instance prediction is wrong.
         # Because of how we're handling masking, masked positions are automatically "correct".
-        correct = predictions.eq(gold_labels).prod(dim=1).float()
+        correct = predictions.eq(gold_labels).prod(dim=1).type(FloatPrecision.dtype)
 
         # Since masked positions are correct, we need to explicitly exclude instance predictions
         # where the entire prediction is masked (because they look "correct").
